+++ test/harness/actions/NetworkController.ts
@@ -92,45 +92,4 @@ export class NetworkController {
     async simulatePeerTimeout(peerIndex: number): Promise<void> {
         return this.disconnectPeer(peerIndex);
     }
-
-    /**
-     * Reconnect a previously disconnected peer
-     */
-    async reconnectPeer(peerIndex: number): Promise<void> {
-        const peer = this.harness.peers[peerIndex];
-        if (!peer) throw new Error(`Peer ${peerIndex} not found`);
-
-        await peer.stateManager.p2pManager.tryOpenConnectionToChannel(
-            this.harness.channelId!.toString()
-        );
-
-        peer.logger.info("Reconnected to P2P network");
-    }
-
-    /**
-     * Disconnect two specific peers from each other
-     */
-    async disconnectPeers(
-        peer1Index: number,
-        peer2Index: number
-    ): Promise<void> {
-        const peer1 = this.harness.peers[peer1Index];
-        const peer2 = this.harness.peers[peer2Index];
-
-        if (!peer1) throw new Error(`Peer ${peer1Index} not found`);
-        if (!peer2) throw new Error(`Peer ${peer2Index} not found`);
-
-        const transport = this.harness.stateQuery.getPeerTransport(
-            peer1Index,
-            peer2Index
-        );
-        if (transport) {
-            peer1.p2pInstance.p2pSigner.p2pManager.disconnectConnection(
-                transport
-            );
-            this.logger.debug(
-                `Disconnected peer ${peer1Index} from peer ${peer2Index}`
-            );
-        }
-    }
 }
