+++ test/harness/blocks/Byzantine.ts
@@ -1,82 +1,85 @@
 import { HarnessBlock } from "./HarnessBlock";
 import { Codec, Type, hash } from "@/utils";
-import { Context } from "./ContextBlocks";
-import { Event } from "./EventBlocks";
 import { SignatureUtils } from "@/utils/SignatureUtils";
 
-/**
- * Tamper helper functions
- * Pure functions that modify dispute objects for testing
- */
-
-/**
- * Tampers with the auditing data hash to make it incorrect
- */
-function tamperAuditingDataHash(dispute: {
-    input: { disputeAuditingDataHash: string };
-}): void {
-    // Tamper: set incorrect auditing data hash (dummy value)
-    dispute.input.disputeAuditingDataHash = hash("0x42");
-}
-
-/**
- * Tampers with the timeout participant address
- */
-function createTamperTimeoutParticipant(
-    wrongParticipantAddress: string,
-    blockHeight: number
-) {
-    return (dispute: {
-        input: { timeout: { participant: string; blockHeight: number } };
-    }): void => {
-        // Tamper: set timeout participant to someone who is NOT next to write
-        dispute.input.timeout.participant = wrongParticipantAddress;
-        dispute.input.timeout.blockHeight = blockHeight;
-    };
-}
-
-/**
- * Tampers BOTH the auditing data hash AND the latest state snapshot hash
- * This causes both commitment check and state proof verification to fail
- */
-function tamperDoubleFault(dispute: {
-    input: { disputeAuditingDataHash: string; latestStateSnapshotHash: string };
-}): void {
-    // Tamper BOTH: auditing data hash (commitment check fails)
-    dispute.input.disputeAuditingDataHash = hash("0x42");
-    // AND: latest state snapshot hash (state proof verification fails)
-    dispute.input.latestStateSnapshotHash = hash("0x43");
-}
-
-/**
- * Tampers the latest state snapshot hash ONLY (commitment stays valid)
- */
-function tamperInvalidStateProof(dispute: {
-    input: { latestStateSnapshotHash: string };
-}): void {
-    // Only tamper the state proof (commitment stays valid)
-    dispute.input.latestStateSnapshotHash = hash("0x42");
-}
-
-/**
- * Tampers with the first milestone's first block to reference an unknown snapshot
- * This makes auditing data reconstruction partial (missing snapshot) and state proof invalid
- */
-function tamperPartialAuditing(dispute: { input: { stateProof: any } }): void {
-    const tamperedStateProof = dispute.input.stateProof;
-    if (
-        tamperedStateProof.milestones.length === 0 ||
-        tamperedStateProof.milestones[0].blockConfirmations.length === 0
-    ) {
-        throw new Error("No milestones to tamper");
+class DisputeTampering {
+    /**
+     * Tampers with the auditing data hash to make it incorrect
+     */
+    static tamperAuditingDataHash(dispute: {
+        input: { disputeAuditingDataHash: string };
+    }): void {
+        // Tamper: set incorrect auditing data hash (dummy value)
+        dispute.input.disputeAuditingDataHash = hash("0x42");
     }
-    const firstBc = tamperedStateProof.milestones[0].blockConfirmations[0];
-    const block = Codec.decode(firstBc.signedBlock.encodedBlock, Type.Block);
 
-    // Tamper: Set to unknown snapshot hash (not stored - makes auditing data partial)
-    block.stateSnapshotHash = hash("0xDEADBEEF");
+    /**
+     * Tampers with the timeout participant address
+     */
+    static createTamperTimeoutParticipant(
+        wrongParticipantAddress: string,
+        blockHeight: number
+    ) {
+        return (dispute: {
+            input: { timeout: { participant: string; blockHeight: number } };
+        }): void => {
+            // Tamper: set timeout participant to someone who is NOT next to write
+            dispute.input.timeout.participant = wrongParticipantAddress;
+            dispute.input.timeout.blockHeight = blockHeight;
+        };
+    }
 
-    firstBc.signedBlock.encodedBlock = Codec.encode(block, Type.Block);
+    /**
+     * Tampers BOTH the auditing data hash AND the latest state snapshot hash
+     * This causes both commitment check and state proof verification to fail
+     */
+    static tamperDoubleFault(dispute: {
+        input: {
+            disputeAuditingDataHash: string;
+            latestStateSnapshotHash: string;
+        };
+    }): void {
+        // Tamper BOTH: auditing data hash (commitment check fails)
+        dispute.input.disputeAuditingDataHash = hash("0x42");
+        // AND: latest state snapshot hash (state proof verification fails)
+        dispute.input.latestStateSnapshotHash = hash("0x43");
+    }
+
+    /**
+     * Tampers the latest state snapshot hash ONLY (commitment stays valid)
+     */
+    static tamperInvalidStateProof(dispute: {
+        input: { latestStateSnapshotHash: string };
+    }): void {
+        // Only tamper the state proof (commitment stays valid)
+        dispute.input.latestStateSnapshotHash = hash("0x42");
+    }
+
+    /**
+     * Tampers with the first milestone's first block to reference an unknown snapshot
+     * This makes auditing data reconstruction partial (missing snapshot) and state proof invalid
+     */
+    static tamperPartialAuditing(dispute: {
+        input: { stateProof: any };
+    }): void {
+        const tamperedStateProof = dispute.input.stateProof;
+        if (
+            tamperedStateProof.milestones.length === 0 ||
+            tamperedStateProof.milestones[0].blockConfirmations.length === 0
+        ) {
+            throw new Error("No milestones to tamper");
+        }
+        const firstBc = tamperedStateProof.milestones[0].blockConfirmations[0];
+        const block = Codec.decode(
+            firstBc.signedBlock.encodedBlock,
+            Type.Block
+        );
+
+        // Tamper: Set to unknown snapshot hash (not stored - makes auditing data partial)
+        block.stateSnapshotHash = hash("0xDEADBEEF");
+
+        firstBc.signedBlock.encodedBlock = Codec.encode(block, Type.Block);
+    }
 }
 
 /**
@@ -140,8 +143,7 @@ export class Byzantine {
             const maliciousPeer = await harness.stateQuery.getNextPeerToWrite();
 
             // Store malicious peer index on harness for use by Event/Assert blocks
-            (harness.context as any).lastMaliciousPeerIndex =
-                maliciousPeer.index;
+            harness.context.lastMaliciousPeerIndex = maliciousPeer.index;
 
             // Submit invalid transition (action only, no waiting)
             await harness.byzantineActions.submitInvalidStateTransitionBlock(
@@ -190,8 +192,7 @@ export class Byzantine {
             const maliciousPeer = await harness.stateQuery.getNextPeerToWrite();
 
             // Store malicious peer index on harness for use by Event/Assert blocks
-            (harness.context as any).lastMaliciousPeerIndex =
-                maliciousPeer.index;
+            harness.context.lastMaliciousPeerIndex = maliciousPeer.index;
 
             // Submit forged inbound message block (action only, no waiting)
             await harness.byzantineActions.submitForgedInboundMessageBlock(
@@ -209,7 +210,7 @@ export class Byzantine {
     static tamperedDisputeAuditingData(peerIndex: number) {
         return Byzantine.postTamperedDisputeWith(
             peerIndex,
-            tamperAuditingDataHash
+            DisputeTampering.tamperAuditingDataHash
         );
     }
 
@@ -229,7 +230,7 @@ export class Byzantine {
                 throw new Error(`Peer ${wrongParticipant} not found`);
             }
 
-            const tamperFn = createTamperTimeoutParticipant(
+            const tamperFn = DisputeTampering.createTamperTimeoutParticipant(
                 notNextPeer.address,
                 blockHeight
             );
@@ -248,7 +249,7 @@ export class Byzantine {
     static tamperedDisputePartialAuditing(peerIndex: number) {
         return Byzantine.postTamperedDisputeWith(
             peerIndex,
-            tamperPartialAuditing
+            DisputeTampering.tamperPartialAuditing
         );
     }
 
@@ -256,7 +257,10 @@ export class Byzantine {
      * Peer posts a tampered dispute with BOTH invalid commitment AND invalid state proof
      */
     static tamperedDisputeDoubleFault(peerIndex: number) {
-        return Byzantine.postTamperedDisputeWith(peerIndex, tamperDoubleFault);
+        return Byzantine.postTamperedDisputeWith(
+            peerIndex,
+            DisputeTampering.tamperDoubleFault
+        );
     }
 
     /**
@@ -265,18 +269,8 @@ export class Byzantine {
     static tamperedDisputeInvalidStateProof(peerIndex: number) {
         return Byzantine.postTamperedDisputeWith(
             peerIndex,
-            tamperInvalidStateProof
+            DisputeTampering.tamperInvalidStateProof
         );
-    }
-
-    /**
-     * Peer skips their turn (doesn't author when expected)
-     */
-    static skipTurn(peerIndex: number) {
-        return new HarnessBlock(async (harness) => {
-            await harness.byzantineActions.skipTurn(peerIndex);
-            return harness;
-        });
     }
 
     /**
@@ -339,134 +333,6 @@ export class Byzantine {
 
             // Store dispute for later assertions
             harness.context.lastTamperedDispute = dispute;
-
-            return harness;
-        });
-    }
-
-    // ========================================
-    // COMPOSED SCENARIOS - High-level attack patterns
-    // ========================================
-    // NOTE: These are meta-blocks that compose multiple primitive blocks.
-    // They represent complete attack scenarios rather than single Byzantine actions.
-
-    /**
-     * META-BLOCK: Trigger invalid state transition from a malicious peer, wait for disputes, and resolve the fork
-     *
-     * This is a high-level composition block that orchestrates multiple primitive blocks
-     * to create a complete fork resolution scenario. It composes:
-     * - Context marking (malicious/honest peers)
-     * - Fork capture
-     * - Byzantine attack (invalid transition)
-     * - Event synchronization (dispute commits, fork change)
-     * - Context update
-     */
-    static createAndResolveFork(options: {
-        maliciousPeerIndex: number;
-        honestPeerIndices?: number[];
-    }) {
-        const { maliciousPeerIndex, honestPeerIndices } = options;
-
-        return new HarnessBlock(async (harness) => {
-            const forkId = harness.activeForkId;
-            if (!forkId) {
-                throw new Error(
-                    "No active fork ID - channel must be opened first"
-                );
-            }
-
-            const totalPeers = harness.peers.length;
-            const honest =
-                honestPeerIndices ||
-                Array.from({ length: totalPeers }, (_, i) => i).filter(
-                    (i) => i !== maliciousPeerIndex
-                );
-
-            // Compose all blocks using HarnessBlock.compose()
-            return HarnessBlock.compose(
-                Context.markMaliciousPeer({
-                    maliciousPeerIndex,
-                    honestPeerIndices
-                }),
-                Event.captureOriginalFork(),
-                Event.reset(),
-                Byzantine.invalidTransitionFrom(maliciousPeerIndex),
-                Event.waitForAllPeers("onDisputeCommitted", honest.length, {
-                    timeoutMs: 5000,
-                    mode: "atLeast"
-                }),
-                Event.waitForForkChange({
-                    timeoutMs: 10000,
-                    honestPeerIndices: honest
-                }),
-                Context.updateActiveFork()
-            ).run(harness);
-        });
-    }
-
-    /**
-     * META-BLOCK: Create and resolve a fork with full settlement control
-     *
-     * This is a high-level composition block that creates an invalid state transition
-     * dispute and waits for fork resolution with configurable timing for dispute commits
-     * and fork settlement.
-     *
-     * Unlike createAndResolveFork(), this provides:
-     * - Control over dispute commit timing (disputesCommittedTimeoutMs)
-     * - Control over fork settlement timing (forkSettleTimeoutMs)
-     * - More lenient dispute commit requirements (some peers may be slow)
-     *
-     */
-    static createAndResolveForkWithSettlement(options: {
-        maliciousPeerIndex: number;
-        honestPeerIndices?: number[];
-        forkSettleTimeoutMs?: number;
-        disputesCommittedTimeoutMs?: number;
-    }) {
-        const {
-            maliciousPeerIndex,
-            honestPeerIndices,
-            forkSettleTimeoutMs = 10000,
-            disputesCommittedTimeoutMs = 5000
-        } = options;
-
-        return new HarnessBlock(async (harness) => {
-            const forkId = harness.activeForkId;
-            if (!forkId) {
-                throw new Error(
-                    "No active fork ID - channel must be opened first"
-                );
-            }
-
-            const totalPeers = harness.peers.length;
-            const honest =
-                honestPeerIndices ||
-                Array.from({ length: totalPeers }, (_, i) => i).filter(
-                    (i) => i !== maliciousPeerIndex
-                );
-
-            // Mark malicious peer context for later blocks
-            harness.context.maliciousPeerIndex = maliciousPeerIndex;
-            harness.context.honestPeerIndices = honest;
-
-            // Use disputeOrchestrator action to handle the complex workflow
-            const result =
-                await harness.disputeOrchestrator.createAndResolveInvalidStateTransitionDispute(
-                    maliciousPeerIndex,
-                    {
-                        forkId,
-                        honestPeerIndices: honest,
-                        forkSettleTimeoutMs,
-                        disputesCommittedTimeoutMs,
-                        resetEventSpies: true,
-                        disputesCommittedMode: "atLeast",
-                        assertMaliciousRemoved: false
-                    }
-                );
-
-            // Update active fork context
-            harness.context.originalForkId = forkId;
-            harness.activeForkId = result.newForkId;
 
             return harness;
         });
@@ -537,16 +403,6 @@ export class Byzantine {
     }
 
     /**
-     * Reconnect all peers (heal all partitions)
-     */
-    static reconnectAll() {
-        return new HarnessBlock(async (harness) => {
-            await harness.networkController.connectAllPeers();
-            return harness;
-        });
-    }
-
-    /**
      * Stub a peer's calldata handler to prevent syncing from on-chain calldata
      */
     static stubCalldataHandler(peerIndex: number) {
@@ -561,7 +417,7 @@ export class Byzantine {
                 eventHandler.onBlockCalldataPosted.bind(eventHandler);
 
             // Store original handler for potential restoration
-            (harness as any)[`peer${peerIndex}OriginalCalldataHandler`] =
+            harness.context[`peer${peerIndex}OriginalCalldataHandler`] =
                 originalHandler;
 
             // Stub with no-op
@@ -583,9 +439,8 @@ export class Byzantine {
                 throw new Error(`Peer ${peerIndex} not found`);
             }
 
-            const originalHandler = (harness as any)[
-                `peer${peerIndex}OriginalCalldataHandler`
-            ];
+            const originalHandler =
+                harness.context[`peer${peerIndex}OriginalCalldataHandler`];
             if (!originalHandler) {
                 throw new Error(
                     `No original calldata handler found for peer ${peerIndex}`
@@ -612,7 +467,7 @@ export class Byzantine {
             const remoteRpc = peer.stateManager.p2pManager.remoteRpc;
 
             // Store original function for potential restoration
-            (harness as any)[`peer${peerIndex}OriginalBroadcast`] =
+            harness.context[`peer${peerIndex}OriginalBroadcast`] =
                 remoteRpc.stateTransitionService.onBlockConfirmation;
 
             // Stub with no-op broadcast
