+++ test/harness/blocks/Scenario.ts
@@ -1,12 +1,13 @@
-import { HarnessBlock, composeBlocks } from "./HarnessBlock";
-import { Lifecycle } from "./LifecycleBlocks";
+import { HarnessBlock } from "./HarnessBlock";
+import { Lifecycle } from "./Lifecycle";
 import { HarnessOptions } from "../core/types";
-import { Assert } from "./AssertBlocks";
-import { Byzantine } from "./ByzantineBlocks";
-import { Transition } from "./TransitionBlocks";
-import { Event } from "./EventBlocks";
-import { Time } from "./TimeBlocks";
-import { Sync } from "./SyncBlocks";
+import { Assert } from "./Assert";
+import { Byzantine } from "./Byzantine";
+import { Transition } from "./Transition";
+import { Event } from "./Event";
+import { Wait } from "./Wait";
+import { Context } from "./Context";
+import { Time } from "./Time";
 
 /**
  * Scenario building blocks for common test patterns
@@ -30,7 +31,7 @@ export class Scenario {
      * Channel configured for timeout testing (short timeouts)
      */
 
-    static timeoutChannel(peerCount: number = 3) {
+    static timeoutSetup(peerCount: number = 3) {
         return HarnessBlock.compose(
             Lifecycle.setup(peerCount, {
                 timeConfig: {
@@ -69,23 +70,9 @@ export class Scenario {
     static advanceState(count: number) {
         return new HarnessBlock(async (harness) => {
             for (let i = 0; i < count; i++) {
-                await harness.transitionActions.submitNext((contract) =>
-                    contract.add(1)
-                );
+                await harness.transitionActions.increment();
             }
 
-            return harness;
-        });
-    }
-
-    /**
-     * Add a specific value to the state
-     */
-    static addValue(value: number) {
-        return new HarnessBlock(async (harness) => {
-            await harness.transitionActions.submitNext((contract) =>
-                contract.add(value)
-            );
             return harness;
         });
     }
@@ -117,17 +104,6 @@ export class Scenario {
     }
 
     /**
-     * @deprecated Use peerWrite() instead (singular)
-     */
-    static peerWrites(options: {
-        peer: number;
-        value?: number;
-        waitForPeers?: number[];
-    }) {
-        return Scenario.peerWrite(options);
-    }
-
-    /**
      * One full round (all peers write once in order)
      */
     static fullRound() {
@@ -135,9 +111,7 @@ export class Scenario {
             const peerCount = harness.peers.length;
 
             for (let i = 0; i < peerCount; i++) {
-                await harness.transitionActions.submitNext((contract) =>
-                    contract.add(1)
-                );
+                await harness.transitionActions.increment();
             }
 
             return harness;
@@ -153,30 +127,12 @@ export class Scenario {
 
             for (let round = 0; round < rounds; round++) {
                 for (let peer = 0; peer < peerCount; peer++) {
-                    await harness.transitionActions.submitNext((contract) =>
-                        contract.add(1)
-                    );
+                    await harness.transitionActions.increment();
                 }
             }
 
             return harness;
         });
-    }
-
-    // ========================================
-    // PRE-COMPOSED SCENARIOS - Common patterns
-    // ========================================
-
-    /**
-     * Channel with one full round completed (all peers wrote once)
-     */
-
-    static oneRound(peerCount: number = 3, options?: HarnessOptions) {
-        return HarnessBlock.compose(
-            Lifecycle.setup(peerCount, options),
-            Lifecycle.openChannel(),
-            Scenario.fullRound()
-        );
     }
 
     // ========================================
@@ -197,7 +153,7 @@ export class Scenario {
         return HarnessBlock.compose(
             Scenario.activeChannel(4, 2, options),
             Assert.allPeersInSync(),
-            Byzantine.createAndResolveFork({ maliciousPeerIndex: 2 })
+            Scenario.forkResolution({ maliciousPeerIndex: 2 })
         );
     }
 
@@ -246,8 +202,7 @@ export class Scenario {
             Assert.allPeersInSync(),
             Event.reset(),
             Byzantine.stubBroadcast(peerIndex),
-            Transition.valid((c) => c.add(value), { waitForSync: false }),
-            Time.wait(1000)
+            Transition.valid((c) => c.add(value), { waitForSync: false })
         );
     }
 
@@ -264,7 +219,7 @@ export class Scenario {
             Scenario.advanceState(initialTransitions),
             Lifecycle.addPeer(),
             Event.waitUntilEventOccurs("onConnection", 5000),
-            Sync.wait({ timeout: 5000 })
+            Wait.untilInSync(undefined, { timeout: 5000 })
         );
     }
 
@@ -325,5 +280,123 @@ export class Scenario {
             Assert.disputeCommitted(),
             Event.reset()
         );
+    }
+
+    // ========================================
+    // FORK RESOLUTION SCENARIOS
+    // ========================================
+
+    /**
+     * Fork resolution scenario: malicious peer creates invalid transition,
+     * honest peers dispute, fork reduces
+     */
+    static forkResolution(options: {
+        maliciousPeerIndex: number;
+        honestPeerIndices?: number[];
+    }) {
+        const { maliciousPeerIndex, honestPeerIndices } = options;
+
+        return new HarnessBlock(async (harness) => {
+            const forkId = harness.activeForkId;
+            if (!forkId) {
+                throw new Error(
+                    "No active fork ID - channel must be opened first"
+                );
+            }
+
+            const totalPeers = harness.peers.length;
+            const honest =
+                honestPeerIndices ||
+                Array.from({ length: totalPeers }, (_, i) => i).filter(
+                    (i) => i !== maliciousPeerIndex
+                );
+
+            // Compose all blocks using HarnessBlock.compose()
+            return HarnessBlock.compose(
+                Context.markMaliciousPeer({
+                    maliciousPeerIndex,
+                    honestPeerIndices
+                }),
+                Event.captureOriginalFork(),
+                Event.reset(),
+                Byzantine.invalidTransitionFrom(maliciousPeerIndex),
+                Event.waitForAllPeers("onDisputeCommitted", honest.length, {
+                    timeoutMs: 5000,
+                    mode: "atLeast"
+                }),
+                Event.waitForForkChange({
+                    timeoutMs: 10000,
+                    honestPeerIndices: honest
+                }),
+                Context.updateActiveFork()
+            ).run(harness);
+        });
+    }
+
+    /**
+     * Fork resolution with full settlement control (longer timeouts)
+     *
+     * This is a high-level composition block that creates an invalid state transition
+     * dispute and waits for fork resolution with configurable timing for dispute commits
+     * and fork settlement.
+     *
+     * Unlike forkResolution(), this provides:
+     * - Control over dispute commit timing (disputesCommittedTimeoutMs)
+     * - Control over fork settlement timing (forkSettleTimeoutMs)
+     * - More lenient dispute commit requirements (some peers may be slow)
+     */
+    static forkResolutionWithSettlement(options: {
+        maliciousPeerIndex: number;
+        honestPeerIndices?: number[];
+        forkSettleTimeoutMs?: number;
+        disputesCommittedTimeoutMs?: number;
+    }) {
+        const {
+            maliciousPeerIndex,
+            honestPeerIndices,
+            forkSettleTimeoutMs = 10000,
+            disputesCommittedTimeoutMs = 5000
+        } = options;
+
+        return new HarnessBlock(async (harness) => {
+            const forkId = harness.activeForkId;
+            if (!forkId) {
+                throw new Error(
+                    "No active fork ID - channel must be opened first"
+                );
+            }
+
+            const totalPeers = harness.peers.length;
+            const honest =
+                honestPeerIndices ||
+                Array.from({ length: totalPeers }, (_, i) => i).filter(
+                    (i) => i !== maliciousPeerIndex
+                );
+
+            // Mark malicious peer context for later blocks
+            harness.context.maliciousPeerIndex = maliciousPeerIndex;
+            harness.context.honestPeerIndices = honest;
+
+            // Use disputeOrchestrator action to handle the complex workflow
+            const result =
+                await harness.disputeOrchestrator.createAndResolveInvalidStateTransitionDispute(
+                    maliciousPeerIndex,
+                    {
+                        forkId,
+                        honestPeerIndices: honest,
+                        forkSettleTimeoutMs,
+                        disputesCommittedTimeoutMs,
+                        resetEventSpies: true,
+                        disputesCommittedMode: "atLeast",
+                        assertMaliciousRemoved: false
+                    }
+                );
+
+            // Update active fork context
+            harness.context.originalForkId = forkId;
+            harness.activeForkId = result.newForkId;
+
+            return harness;
+        });
     }
 }
